logo component test fiel
